Shuffle things in to d.ts and utils files
